+ init: switch to asyncio event model using kernel input char devices
we previously had the comment about /sys/kernel/irq/N/per_cpu_count:

> inotify doesn't notice modifies, and select returns immediately, so
> we're forced to parse and count at intervals.  no event mechanism known

but this is no longer true, as we discovered /dev/input/event* char
devices are not exclusive to Xorg as we originally thought, but actually
can be listened by multiple readers, using blocking reads (see my answer
https://unix.stackexchange.com/questions/590831).

this solves our problem that we didn't want to poll, but rather use an
event loop, since most of the time we'll be asleep and don't want to
wake up at intervals just to read the interrupt file and go back to
sleep.

it also solves another problem of not knowing the right irq number, ie
from removed comment:

> touchpad: 144 (also 14, 17 but 144 is closest: DLL096D:01, may have to
> parse /proc/interrupts and look up this symbolic name because the
> allocated interrupt might vary each boot)

it does in fact vary each boot and there's no way to even look it up in
/proc/interrupts because there's more than one xhci interrupt that are
varying over time.  there's probably no way to do it programmatically.
but fortunately with /dev/input/by-path the kernel is doing this work
for us and there's a glob pattern we can use to identify all the
keyboard and mouse devices, yay! with luck, in future we can use inotify
on this directory to check when input devices have changed, such as when
unplugging a keyboard.

right now this is only a skeleton that establishes the event loop, opens
files and spawns a thread to monitor them (because they are blocking
reads on the event devices, and we're using python asyncio, which only
has one thread doing the coroutines in cooperative multitasking), which
prints them out.  we still have to make an event processing coroutine
and give it to them from the threads using a thread-safe queue structure
of some sort.  but this proof of concept is at least as far as the
previous code was.

this was a wholesale replace with totally different code, and we're
still at proof of concept stage, so there was no sense breaking up the
patch into logical commits