= imports: import only needed stuff from argparse, not whole module
to keep everything like the other ones.  mostly we prefer to use it like
'from foo import bar' to avoid having to prepend to namespace everywhere
in the code ('import par, foo.bar'), which makes code lines shorter

as far as speed, 'from foo import bar' makes 'bar()' faster, because it
avoids a namespace lookup, but the import itself is slower
(https://stackoverflow.com/questions/3591962), so it depends on how many
times the call is done which way is faster.  but either way is trivially
small amount of time and not really worth considering.  indeed, that's
not the reason we do it, but more for shorter code (as above).  opinions
will differ on whether explicit namespace use in module calls make the
code less or more readable