- queue: print exception from read() since it wouldn't otherwise
because asyncio.to_thread() uses a ThreadPoolExecutor, which makes
"daemon threads," they won't be cleaned up by the main thread on
exception (basically they're untracked by the main thread, as if they
didn't exist), and thus, program will die without thread exceptions
being printed on their own.  we want to see when exceptions happen
though, so we can write code later to handle them.  probably they will
except when the keyboard is removed for example, so we'll want to make
this a more specific catch later when we implement code to handle input
device changes, but for now it will suffice to just print the exception