+ process: SimpleQueue in queue_events() and .get() process_events()
so far we wrote code to read events from input devices, discard the ones
we don't care about, and print the ones we collected.  but we ultimately
need to actually "do things" with these events, so now we have added a
process_events() thread which .get()s from a thread-safe SimpleQueue
instance (there are potentially multiple keyboard or mouse event
threads, so we need locking because counter increment is not atomic in
python) and prints the events that come in.  likewise queue_events()
is modified to .put() the events onto the queue for processing.

because we will use a different threshold for ignoring bogus keyboard vs
mouse events, we'll need either two queues, or one queue with a "type"
so the events can be handled accordingly.  we chose the latter to
simplify the code because queue .get() is a blocking call and this means
we would have needed a separate thread for keyboard and mouse
processing.