+ queue: also store timestamp of last seen packet and print it
when we see packets, we have to figure out how much time has elapsed
since the last packet was seen.  rather than getting the current time,
we can just keep the one that the kernel gave us, as the input events
have a struct timeval attached.  fortunately this is pretty much the
same format as time.time(), ie a long seconds since epoch, and a long
which is the number of microseconds.  time.time() is a float that's the
former with the decimal as the latter, so it's an easy conversion.

this is a naive implementation that only stores the last seen timestamp.
in reality we need to get a rate and say "have we seen so many in the
last interval that we consider events for?" so we will have to store
some kind of rotating buffer of them ultimately, but in order to test
that the events were being procesed into times correctly and ferried
from the kernel events, this is a good first pass