+ queue: watch for SYN_DROPPED although it shouldn't happen
previously we were just ignoring SYN_DROPPED on the grounds that it
shouldn't ever happen.  it would mean we couldn't read() events out of
the kernel buffer as fast as they were being put into there and kernel
filled it up, but this should basically never happen because we're doing
very little packet processing and have four cores to get scheduled on

instead of ignoring, we'll see if this actually happens in practice by
bombing out should it occur.  do not anticipate this ever happening but
it's just a single value test each loop, shouldn't cause performance
issues