+ process: preseed event ring with time from unidle_s ago
the program starts in the idle state.  if we initialize the circular
buffer with the program start time, that makes the very first time
comparison (which will be from the key release after pressing "enter" on
the command line, if started interactively) always be a very small
number and it will immediately end up calling unidle().  instead if we
seed it with a value equal to program start time minus the number of
seconds in the assessment interval, it'll always be infinitesimally more
idle than the unidle interval and shouldn't trigger the needless state
change to unidle