+ process: use timeout on event gets to go idle after timeout
we track a global idle state and, if we're not currently idle, we can
make a timeout-bounded call to read off the event queue; upon expiry,
the interval has passed and we'll be considered "idle" (ie "afk").

we don't require a lower-than-threshold number of events per time like
we require a higher-than-threshold number of events for coming back to
the keyboard ("unidle"), because it's only for a small window of time
that it's even possible to happen, and the user is actually at the
keyboard, so accidentally *not* pressing keys would be rare.  we can
govern this instead by just incrementing the threshold during which we
can't see any keypresses, if we find it happening, and this should be a
good enough adjustment for going afk.  we'll see how it works out in
practice

this will cause a minor version bumb since it's a new capability and
ever-closer to the initial feature-complete version