+ process: break out actions on idle/unidle to full functions
for now these will just print but the whole point of the program --
beyond keeping a log for later reference when needing to enter tracking
fudges when we forget to stop the clock -- is to actually do something
on afk/back.  these functions will end up implementing that under
control of the user, probably in an rcfile that defines what to do.