/ queue: do queue insert each input packet instead of each read
we can sometimes get multiple SYN_REPORT in one read() buffer, so we
need to handle it based on input packets and do the queue work each
loop.  this was previously unknown, and was suspected not to happen.  it
doesn't happen often (as determined by debug prints), but does happen