+ watch: monitor for input dev changes with inotify in by-path/
udev creates and deletes symlinks in /dev/input/by-path/ every time a
device is added or removed.  unfortunately this will add dependency on
inotify library but we're already using yaml library for rcfile, so hrm

for now we just print when the type is a create or delete event.  in the
future we should check that the filename matches the pattern of a device
we are interested in