+ log: add log() to replace print() and make err() wrap log()
err() logs to stderr.  we don't want to just use print() for when we
want stdout, because err() might end up doing other things (like maybe
always printing a timestamp), so we don't want to do that in two places.

instead we make log() the new name we'll use where we would have used
print() (which for now, just prints anyways).  then we'll make err()
call log() but also pass stderr as the file kwarg.

note that this use of 'file' kwarg -- which is implementation-specific
and relies on print() being the backend -- will have to use a different
method if/when we ever make log() do something else (like write to files
or syslog or something user-configurable)