- watch: make inotify block forever, are already in separate thread
inotify is using epoll() with a timeout of 1s by default.  this is
intended for single threaded programs to "do stuff" in-between polls.
since we already have different threads we can block on i/o forever and
yield the processor handily, so we have no need for these 1s interval
interruptions.

there's no way to disable them apparently so we just give a high value
that will never happen (by experimentation 20 was the highest shift in
the acceptable range for the function, this means every 9 days we will
need an extra system call, saving 777600 syscalls compared to doing it
each second for the same duration.