+ monitor: if thread crashes, cancel other pending and dump stack
long has debugging been difficult because we don't get stack traces from
the threads without catching it in the thread at the site of use and
printing it.  this makes development annoying.  while we'll probably
leave those catches in, this enables a catch-all in the monitor
itself... the first uncaught exception in any thread *should* safely be
logged with this method (tested to also work with a chained one)

note that canceling doesn't do anything if the canceled function blocks
(it won't unblock in order to cancel).  most of our other threads would
be blocking themselves, so this is largely perfunctory, but seems to
work out fine, at least it might save some error messages while waiting
for execv() to whisk away life itself