= global: section globals we may need to share with re-exec'd self
put these in their own section and make clear Lock() is not needed.  we
are collecting variables that we might have to pass to a successor
process after hangup.  these will be global variables and we'll have to
replace the init of them also, for now we are putting the declarations
together to start with