+ monitor: gain capability to handle device events without restart
when input devices change, we were previously restarting the daemon
completely (ie, execv()) to reopen new set of devices correctly, so like
just starting over when device removal was detected (EBADF or ENODEV on
the read fd).

plain restart loses state, such as idle state and last seen input packet
timestamps.  we could have tried to transfer this to the newly-exec'd
child via some shared memory, a pipe or command line arguments, but this
turns out to be hard to get right and seems not worth the effort -- if
we're going to all this trouble, we might as well just maintain enough
state internally to restart and then we don't have to pass any state to
the re-exec()ed program.

so this patch handles it internally by making the input event loop
restartable, in that FIRST_EXCEPTION we can generate ourselves when we
want the event loop to restart.  then we can make the queue_events(fd)
thread for disappearing input device also to go away, while spawning
other new ones based on [the now-globally-maintained] eventfds list, and
continue to wait() on the new set of awaitable threads which should be
set to the fixed ones plus one per fd we're monitoring

we keep process_events() and watch_devices() as the fixed persistent
task threads that we never intend to remove, but then let each of the fd
queue threads (one per input device) handle disappearing file
descriptors (as determined by read() error) by throwing a custom
exception IdledRescan

IdledRescan will let the monitor thread re-loop, and in the meantime
the new abort_queues() is run.  this function aborts all the queues
threads by moving through all the open file descriptors as indices into
a list of thread IDs which can be given to pthread_kill().  we send the
custom SIGUSR1 signal and rig it to do nothing, but in handling the
signal will succeed in returning from the otherwise blocked queue
read(fd).  this is the only way to "wake up" the thread and have an
orderly exit, unfortunately read() on a close()ed file descriptor
doesn't seem to return immediately when the fd becomes invalid, it seems
to wait until some other inputs have come.  in such a way we can ensure
that we signal all extant fd threads, and proceed to reopen anew
following rescan of devices.

finally, we also empower the watch_devices() thread to respond to new
device paths being added and removed by the kernel, by initiating the
same function abort_queues(), causing a rescan.  this would be the only
thing that could handle a new device, but duplicates the functionality
of missing fd