+ rcfile: replace yaml with sectionless configparser ast.literal_eval
yaml requires a separate library, which we want to avoid.  too bad,
since yaml would be the best, but we don't want this daemon to have any
dependencies at this time, it should be a standalone script

note that this change required no changes to the rcfile itself.  without
ast.literal_eval, everything would be strings, and then we would have to
"know" the type of the object in the parser, which defeats the purpose

it will require quoting strings though, when we want to pass a shell
script command as something to execute if we add warning that idle will
happen soon (in form of something to pass to os.system()