/ log: flush prints, else logs not flushed when outputting to pipe
printing won't actually emit anything unless we flush it.  this happens
when stdout is a pipe, presumably BUFSIZ or PIPEBUF is used.  my guess
is, python detects if stdout is a tty and if so flushes output each call
(or each line?), but if a pipe, it will only flush when the buffer is
full